zero_sup_threshold changed to noise multiplication factor.

+++ src/hexsample/display.py
@@ -78,7 +78,7 @@ class HexagonalGridDisplay:
     """Display for an HexagonalGrid object.
     """
 
-    def __init__(self, grid: HexagonalGrid, zero_sup_threshold: int = 0,
+    def __init__(self, grid: HexagonalGrid, zero_sup_threshold: float = 0.,
                  **kwargs) -> None:
         """Constructor.
         """
@@ -152,7 +152,7 @@ class HexagonalGridDisplay:
 
     def draw_digi_event_rectangular(self, event: DigiEventRectangular,
         offset: Tuple[float, float] = (0., 0.),
-        indices: bool = True, padding: bool = True, zero_sup_threshold: float = 0,
+        indices: bool = True, padding: bool = True, zero_sup_threshold: float = 0.,
         values: bool = True, **kwargs) -> HexagonCollection:
         """Draw an actual event int the parent hexagonal grid.
 
@@ -170,7 +170,7 @@ class HexagonalGridDisplay:
         return collection
 
     def draw_digi_event_circular(self, event: DigiEventCircular,
-        offset: Tuple[float, float] = (0., 0.), zero_sup_threshold: float = 0,
+        offset: Tuple[float, float] = (0., 0.), zero_sup_threshold: float = 0.,
         values: bool = True, **kwargs) -> HexagonCollection:
         """Display a digi event with circular readout.
         """
@@ -244,7 +244,7 @@ class HexagonalGridDisplay:
 
 class EventDisplay(HexagonalGridDisplay):
 
-    def __init__(self, input_file, grid: HexagonalGrid, zero_sup_threshold: int = 0, **kwargs):
+    def __init__(self, input_file, grid: HexagonalGrid, zero_sup_threshold: float = 0., **kwargs):
         super().__init__(grid, zero_sup_threshold, **kwargs)
         self._input_file = input_file
         self.event_id = kwargs.get("event_id", 0)
@@ -292,7 +292,7 @@ class EventDisplay(HexagonalGridDisplay):
     def current_zero_sup_threshold(self) -> int:
         """Convenience method to get the current zero suppression threshold.
         """
-        return int(self.zero_sup_text_box.text)
+        return float(self.zero_sup_text_box.text)
 
     def _draw(self, event: DigiEventBase) -> None:
         """Complete draw method for an event.
